ghostscript-9.50 updates
Since version 9.50 GS defaults to -dSAFER and produces a warning
messages if -dWRITESYSTEMDICT is used, which ends up in the output
file after the psfrem stage, messing it up.

The proper way to get WRITESYSTEMDICT now is to specify -dNOSAFER
explicitly and do .setsafer once we're done with the systemdict.

The patch also fixes a minor issue with psfrem not truncating its
output.